+++ src/Test/ParseMetadata.hs
@@ -252,6 +252,7 @@ tests = [
 
     checkWriteFail "compile mode" $ ExecuteTests { etInclude = [] },
     checkWriteFail "compile mode" $ CompileRecompile,
+    checkWriteFail "compile mode" $ CompileRecompileRecursive,
     checkWriteFail "compile mode" $ CreateTemplates,
     checkWriteFail "compile mode" $ CompileUnspecified
   ]