+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## 0.4.1.0  -- 2020-05-05
 
+* **[new]** Adds a compiler mode (`-R`) to recursively recompile modules.
+
 * **[new]** Improves thread-safety of internal code so that thread support can
   be safely added later. (Probably via a library.)